+++ netx/oldhttptransport/bodytracer.go
@@ -1,4 +1,4 @@
-package httptransport
+package oldhttptransport
 
 import (
 	"io"
