+++ netx/oldhttptransport/tracetripper.go
@@ -1,4 +1,4 @@
-package httptransport
+package oldhttptransport
 
 import (
 	"bytes"
